Meta: GitHub Boards Automation (#1549)
Adds a few automation to the GitHub project board to help manage the incoming reports:

* Moving a card to "Not doing" column closes it and removes it from the board
* Moving a card to "Future work" removes it from the board
* Moving a card to "Need Author's reply" adds a "Need Author's reply" label
* When the author replies, the label gets removed and the card is moved back to inbox

## Implementation details

Every 10 minutes, a PHP script talks to the GitHub API and looks for cards matching the automation criteria and applies any required updated.

Unfortunately this is the simplest setup I've discovered. Reacting to WebHooks or relying on the regular GitHub workflows would be tedious to setup and maintain for 4 repositories and, separately, for the WordPress organization (to get updates about the project board).

## Testing Instructions (or ideally a Blueprint)

As with any GitHub workflow – just merge and test it live.