Autopublish npm packages every week (#1542)
## Motivation for the change, related issues

Publishes Playground packages every week to enable integrators to stay
on top of the latest changes.

The manual "publish when needed" approach we've used so far is no longer
suitable. It was based on an assumption that the most interested
developers will keep up with the project and ask for a new release when
needed. In reality people tend to keep up with the npm packages, not the
GitHub repo. A few developers reached out and directly asked for more
frequent releases.

This PR ensures there's a patch release every Monday. Let's see how well
it works and continue iterating on the approach should any problems
arise. In the future, we might want to explore formal release trains,
beta periods, and such, but today I believe we can get away with a much
leaner approach.