Blueprints: Add resetData step to documentation. (#1658)
`resetData` step for blueprints exists and it's the recommended way to
reest data IDs. However, this step doesn't appear in the list of steps
available for blueprints in the docs.

This PR adds that step to the documentation by removing the internal and
private annotations.

## Testing instructions

Run `nx build docs-site` and `npm run dev:docs` locally. Go to
http://localhost:3000/wordpress-playground/blueprints-api/steps#ResetDataStep
and confirm the `resetData` step is there.

cc @juanmaguitar

Fixes #1653