Add ESLint rules to limit comment width (#1613)
Long doc comments are hard to read and it's easy to write long comments.

We need an automated way of checking for the comment length.

This PR adds
[eslint-plugin-comment-length](https://www.npmjs.com/package/eslint-plugin-comment-length?activeTab=readme)
to implement comment max-len `eslint` rules.

The errors aren't fixed to avoid creating issues with this PR, but they
will show up as warnings, so we can fix them in the future.

- CI