Fix sqlite-database-integration rename fatal (#1695)
## Motivation for the change, related issues

This is a fix for a fatal that occurs when loading the latest
sqlite-database-integration version from its `develop` branch in
Playground.

Playground initialization renames a temporary
sqlite-database-integration plugin directory based on the previous
branch name of main like "sqlite-database-integration-main", but now
that temporary directory is "sqlite-database-integration-develop".

## Implementation details

We check for the existence of the old temp directory and use that if it
exists. Otherwise, we rename based on the new temp directory name.

## Testing Instructions (or ideally a Blueprint)

Tested locally via `npm run dev`.