WebApp Redesign: Interact with sites list via redux (#1679)
## Motivation for the change, related issues

This is a PR to start exploring site storage. In order to start from a
concrete place, this PR starts by setting up the site manager sidebar to
interact with the sites list via redux.

Related to #1659

## Implementation details

This PR adds a `site-storage` module that provides functions for adding,
removing, and listing sites. It currently only supports writing to sites
stored in OPFS.

Today, Safari only appears to support writing to OPFS files from worker
threads, so this PR adds a `site-storage-metadata-worker.ts` module that
the UI thread can spawn to write site metadata to OPFS.

This PR adds a `siteListing` property to our web app's redux state. It
looks like:
```ts
{
    status: SiteListingStatus;
    sites: SiteInfo[];
}
```

`SiteListingStatus` can reflect whether the listing is loading, loaded,
or in an error state.

The site-manager-sidebar has been updated to select sites state from
redux and interact with the sites list via redux actions. Currently, the
loading status is ignored, but we should show it to the user in a
follow-up PR.

## Testing Instructions (or ideally a Blueprint)

Run `npm run dev` and interact with the sites list in Chrome and Safari.
(Unfortunately, Firefox does not yet support loading Service Worker
modules, and that feature is required for our current dev setup)