Use a separate iframe for every Playground site
Resolves the "Playground already booted" error that occurs when
`startPlaygroundWeb` is called more than once on the same iframe.

 ## Implementation

This implementation recreates the iframe every time the site slug
changes. This ensures that all postMessage calls related to the new site
are not processed in the context of a previously booted worker.

 ## Remaining work

I am not happy with the data flows in this PR. The `Main` component
calls `useBootPlayground`, gets an iframeRef, and passes it down to the
`JustViewport` component via props, where it is used in conjunction with
the `siteSlug` extracted from the redux state. This roundabout logic
will bite us.

Let's implement a straightforward data flow instead and either:

* Call `bootPlaygroundWeb()` in `JustViewport` and handle the iframeRef
  right there, or...
* Manage the playground's `iframeRef` through redux state instead of props.

I want to set the stage for keeping the "closed" iframes around for some
more time to enable quick, boot-less switching between sites.

cc @brandonpayton