Extend allowable resources available via WordPress/WordPress.
This extends the allowable resources available via the `wordpress-branch` query string parameter to include all available resources:

* the HEAD of any branch in the form x.x or x.x-branch,
* a specific tag in the form x.x.x, normalizing the value for major releases,
* a specific commit in the form [a-z0-9]{7,40}

Determining the location of the zip file is improved to allow for the dynamic nature of the final URLs. A request is first made to the offical URL to retrieve the location header for the final zip file.