Add initial site info view (#1701)
## Motivation for the change, related issues

This is a PR to add the site info view for the Playground web app
redesign.

Today, we can add sites to browser storage through the UI, but there is
no UI for deleting sites. That comes with the site info view.

## Implementation details

We rely on the `useNavigator()` WordPress hook for managing the current
state. The site slug is partially passed through the URL and partially
through the `Main` component state – we'll need to make it consistent
and potentially store it in Redux.

## Follow-up work

* Get rid of the `Navigator` component and use a custom URL state
management. Navigator isn't helping us. It doesn't support displaying
multiple views at the same time, it brings in custom animations that
conflict with the design we're building, and it doesn't add that much
over a simple state management layer.
* Store the site slug in redux or in the navigation state. Right now the
first time the sidebar is opened, no site is selected.
* Implement the missing mobile interactions, e.g. opening a site.
* Site actions (download as ZIP etc.)
* Reliable site switching ("Playground already booted" error. It may be
out of scope for this PR.)
* Sourcing admin username and password. I don't think we can reliably
provide that information, e.g. what should we do when a site is
imported? Perhaps instead we could offer a special Playground password
that always works for the admin user.
* Attributing logs to each site. Perhaps storing them in OPFS/locally.
Right now we're treating logs as a global list and are displaying the
same list of logs for every site.

## Testing Instructions (or ideally a Blueprint)

Reproduce my test run, try different ways of breaking the site:


https://github.com/user-attachments/assets/b7cb1bab-e85d-4eaa-b5d1-6073c65a2b3e


https://github.com/user-attachments/assets/3b56c2cc-0119-481e-b8af-83cdb36980b4

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>