Add support for handling symlinks in the request handler (#1724)
## Motivation for the change, related issues

The PHPRequestHandler class currently does not work when requesting
symlinks directly.

E.g. a symlink is created on the virtual filesystem and you then try to
request that symlink directly via a HTTP request. The logic is such that
it only handles files and directories and returns with a 404 for valid
symlinks.

To be clear: this only affects directly requesting files or directories
that are symlinks. Requesting files inside a symlinked directory works
fine. Additionally, PHP follows symlinks as expected so this only
affects the request handler.

## Implementation details

The PHP isFile() and isDir() methods were updated to follow symlinks, and that allows PHPRequestHandler to serve symlinked resources with no code changes.

I also added two new methods to the PHP class `isSymlink`
and `readlink` that let us determine if a path is a symlink and what the
target of that symlink is.

After feedback, I've added the `realpath` helper method. It will follow
the symlinks and return the final path. I've also added a way to deal
with paths where the symlink appears in the middle, e.g.
`/some/path/symlinked-directory/foo.php`

## Testing Instructions (or ideally a Blueprint)

---------

Co-authored-by: Brandon Payton <brandon@happycode.net>