Add Playwright tests for UI redesign changes (#1769)
## Motivation for the change, related issues

https://github.com/WordPress/wordpress-playground/pull/1731 introduced
some breaking changes.
Long-term we want to move to Playwright from Cypress tests. This PR is a
first step that introduces Playwright and implements tests broken by the
UI redesign changes in #1731.

## Implementation details

This PR adds support for Playwright as an nx command and as a CI check. 

It keeps all working Cypress tests and replaces broken tests with
Playwright tests.

To make writing of Playwright tests we added fixtures that ensure
WordPress is loaded and can access the content from the WordPress
iframe.

Playwright has issues with JS modules, so this PR changes the type of
the Website project to `commonjs`.

## Testing Instructions (or ideally a Blueprint)

- CI

---------

Co-authored-by: Brandon Payton <brandon@happycode.net>
Co-authored-by: Bart Kalisz <bartlomiej.kalisz@gmail.com>