Fix address bar trailing slash bug
Before this change, adding query params to a dir URI
in the address bar would lead to the address bar adding
a trailing slash after the query params. Since the PHP
request handler already redirects dir URIs to include
a trailing slash, we can remove this code and stop
the bug from happening.