Fix go-to-site menu items to reveal site view (#1833)
## Motivation for the change, related issues

There is a dead zone where the Playground view is hidden by the
Playground Manager yet the "Homepage" and "WP Admin" links do not review
the Playground view when navigating to those pages. This page fixes
that.

Thanks for @chriszarate for reporting!

## Implementation details

We use `useMediaQuery()` in a parent component to inform the site info
child component that the Playground view is expected to be hidden. Then,
we hide the Playground Manager when navigating if the Playground view is
not already shown.

## Testing Instructions (or ideally a Blueprint)

- `npm run dev`
- Navigate to the local dev site
- Open Playground Manager
- Resize the window to look like the following
![Screenshot 2024-09-30 at 5 32
56 PM](https://github.com/user-attachments/assets/1d7b1898-ad43-4438-ac29-419835b62cb5)
- Click "Homepage" in the hamburger menu and confirm that the Playground
view is shown and the homepage displayed.
- Reopen the Playground Manager, click "WP Admin" in the hamburger menu,
and confirm that the Playground view is shown and the WP Admin Dashboard
displayed.