[Remote] Remove the "light" PHP.wasm bundle. Only offer the "kitchen-sink" one
Maintaining the "light" PHP.wasm bundle is a burden (#1848), especially with the
JSPI support on the horizon (#1339). Only 1.3% of Playgrounds are loaded
with the `light` extension bundle (see #1848).

Let's thus simplify the project maintenance and remove the "light" bundle. After this PR, every Playground will be loaded with the "kitchen-sink" extension bundle.
This PR also drops the name "kitchen-sink". From now on, it's just
"php",

 ## Testing instructions

* Confirm Playground continues to work in the local dev env
* Confirm the CI is green – it would perform a thorough test run of many user flows.

Closes #1848