Allow installing Plugins/Themes into an arbitrary folder (#1803)
This allows for installing a user-generated ZIP file into a specific
theme/plugin path.

See #1802 

## Implementation details

`InstallAssetOptions` has a new param, `targetSlug` (to match it's
existing `targetPath`) which if set the ultimate install path is
`targetPath/targetSlug`. If the param is not set, it defaults to
guessing from the ZIP filename and ZIP contents as current.

`InstallPluginOptions` and `InstallThemeOptions` both gain the same
parameter, and pass it through if set.

`InstallThemeOptions` was previously not actually in use, this fixes
that, bringing it inline with `InstallPluginOptions`.

Tests are a bit of a weird mess, as the testing structure doesn't
cleanup between iteration runs, and the helper methods are different
between each section. `zipFiles` helper from the plugin tests should
probably be a common helper instead.

The tests should show the current "issue" with 6b3b5a5 failing, and the
latest changeset resolving that error.

## Testing Instructions (or ideally a Blueprint)

```
{
  "landingPage": "/wp-admin/admin.php?page=example",
  "steps": [
    {
      "step": "login",
      "username": "admin",
      "password": "password"
    },
    {
      "step": "installPlugin",
      "pluginZipFile": {
        "resource": "wordpress.org/plugins",
        "slug": "hello-dolly"
      },
      "options": {
        "targetSlug": "hello-example-plugin"
      }
    },
    {
      "step": "writeFile",
      "path": "/wordpress/wp-content/mu-plugins/example.php",
      "data": "<?php add_action('admin_menu', function(){ add_menu_page('Example','Example',0,'example',function(){ echo '<pre>'; print_r( glob( WP_PLUGIN_DIR . '/*' ) ); echo '</pre>'; }); } );"
    }
  ]
}
```

This should display a admin page with the following output:
```
Array
(
    [0] => /wordpress/wp-content/plugins/hello-example-plugin
    [1] => /wordpress/wp-content/plugins/hello.php
    [2] => /wordpress/wp-content/plugins/index.php
)
```